+++ client/global.go
@@ -3,10 +3,10 @@ package client
 import (
 	"bytes"
 	"crypto/md5"
-	"encoding/base64"
 	"encoding/hex"
 	"encoding/xml"
 	"fmt"
+	"github.com/Mrs4s/MiraiGo/client/pb/oidb"
 	"math/rand"
 	"net"
 	"sort"
@@ -18,7 +18,6 @@ import (
 	"github.com/Mrs4s/MiraiGo/binary/jce"
 	devinfo "github.com/Mrs4s/MiraiGo/client/pb"
 	"github.com/Mrs4s/MiraiGo/client/pb/msg"
-	"github.com/Mrs4s/MiraiGo/client/pb/oidb"
 	"github.com/Mrs4s/MiraiGo/message"
 	"github.com/Mrs4s/MiraiGo/utils"
 	"github.com/pkg/errors"
@@ -382,7 +381,7 @@ func GenIMEI() string {
 		sum += toAdd
 		final.WriteString(fmt.Sprintf("%d", toAdd)) // and even printing them here!
 	}
-	var ctrlDigit int = (sum * 9) % 10 // calculating the control digit
+	var ctrlDigit = (sum * 9) % 10 // calculating the control digit
 	final.WriteString(fmt.Sprintf("%d", ctrlDigit))
 	return final.String()
 }
@@ -507,147 +506,6 @@ func (c *QQClient) parseTempMessage(msg *msg.Message) *message.TempMessage {
 	}
 }
 
-func (c *QQClient) parseGroupMessage(m *msg.Message) *message.GroupMessage {
-	group := c.FindGroup(m.Head.GroupInfo.GetGroupCode())
-	if group == nil {
-		c.Debug("sync group %v.", m.Head.GroupInfo.GetGroupCode())
-		info, err := c.GetGroupInfo(m.Head.GroupInfo.GetGroupCode())
-		if err != nil {
-			c.Error("error to sync group %v : %+v", m.Head.GroupInfo.GetGroupCode(), err)
-			return nil
-		}
-		group = info
-		c.GroupList = append(c.GroupList, info)
-	}
-	if len(group.Members) == 0 {
-		mem, err := c.GetGroupMembers(group)
-		if err != nil {
-			c.Error("error to sync group %v member : %+v", m.Head.GroupInfo.GroupCode, err)
-			return nil
-		}
-		group.Members = mem
-	}
-	var anonInfo *msg.AnonymousGroupMessage
-	for _, e := range m.Body.RichText.Elems {
-		if e.AnonGroupMsg != nil {
-			anonInfo = e.AnonGroupMsg
-		}
-	}
-	var sender *message.Sender
-	if anonInfo != nil {
-		sender = &message.Sender{
-			Uin:      80000000,
-			Nickname: string(anonInfo.AnonNick),
-			AnonymousInfo: &message.AnonymousInfo{
-				AnonymousId:   base64.StdEncoding.EncodeToString(anonInfo.AnonId),
-				AnonymousNick: string(anonInfo.AnonNick),
-			},
-			IsFriend: false,
-		}
-	} else {
-		mem := group.FindMember(m.Head.GetFromUin())
-		if mem == nil {
-			group.Update(func(_ *GroupInfo) {
-				if mem = group.FindMemberWithoutLock(m.Head.GetFromUin()); mem != nil {
-					return
-				}
-				info, _ := c.getMemberInfo(group.Code, m.Head.GetFromUin())
-				if info == nil {
-					return
-				}
-				mem = info
-				group.Members = append(group.Members, mem)
-				go c.dispatchNewMemberEvent(&MemberJoinGroupEvent{
-					Group:  group,
-					Member: info,
-				})
-			})
-			if mem == nil {
-				return nil
-			}
-		}
-		sender = &message.Sender{
-			Uin:      mem.Uin,
-			Nickname: mem.Nickname,
-			CardName: mem.CardName,
-			IsFriend: c.FindFriend(mem.Uin) != nil,
-		}
-	}
-	var g *message.GroupMessage
-	g = &message.GroupMessage{
-		Id:             m.Head.GetMsgSeq(),
-		GroupCode:      group.Code,
-		GroupName:      string(m.Head.GroupInfo.GroupName),
-		Sender:         sender,
-		Time:           m.Head.GetMsgTime(),
-		Elements:       message.ParseMessageElems(m.Body.RichText.Elems),
-		OriginalObject: m,
-	}
-	var extInfo *msg.ExtraInfo
-	// pre parse
-	for _, elem := range m.Body.RichText.Elems {
-		// is rich long msg
-		if elem.GeneralFlags != nil && elem.GeneralFlags.GetLongTextResid() != "" {
-			if f := c.GetForwardMessage(elem.GeneralFlags.GetLongTextResid()); f != nil && len(f.Nodes) == 1 {
-				g = &message.GroupMessage{
-					Id:             m.Head.GetMsgSeq(),
-					GroupCode:      group.Code,
-					GroupName:      string(m.Head.GroupInfo.GroupName),
-					Sender:         sender,
-					Time:           m.Head.GetMsgTime(),
-					Elements:       f.Nodes[0].Message,
-					OriginalObject: m,
-				}
-			}
-		}
-		if elem.ExtraInfo != nil {
-			extInfo = elem.ExtraInfo
-		}
-	}
-	if !sender.IsAnonymous() {
-		mem := group.FindMember(m.Head.GetFromUin())
-		groupCard := m.Head.GroupInfo.GetGroupCard()
-		if extInfo != nil && len(extInfo.GroupCard) > 0 && extInfo.GroupCard[0] == 0x0A {
-			buf := oidb.D8FCCommCardNameBuf{}
-			if err := proto.Unmarshal(extInfo.GroupCard, &buf); err == nil && len(buf.RichCardName) > 0 {
-				groupCard = ""
-				for _, e := range buf.RichCardName {
-					groupCard += string(e.Text)
-				}
-			}
-		}
-		if m.Head.GroupInfo != nil && groupCard != "" && mem.CardName != groupCard {
-			old := mem.CardName
-			if mem.Nickname == groupCard {
-				mem.CardName = ""
-			} else {
-				mem.CardName = groupCard
-			}
-			if old != mem.CardName {
-				go c.dispatchMemberCardUpdatedEvent(&MemberCardUpdatedEvent{
-					Group:   group,
-					OldCard: old,
-					Member:  mem,
-				})
-			}
-		}
-	}
-	if m.Body.RichText.Ptt != nil {
-		g.Elements = []message.IMessageElement{
-			&message.VoiceElement{
-				Name: m.Body.RichText.Ptt.GetFileName(),
-				Md5:  m.Body.RichText.Ptt.FileMd5,
-				Size: m.Body.RichText.Ptt.GetFileSize(),
-				Url:  "http://grouptalk.c2c.qq.com" + string(m.Body.RichText.Ptt.DownPara),
-			},
-		}
-	}
-	if m.Body.RichText.Attr != nil {
-		g.InternalId = m.Body.RichText.Attr.GetRandom()
-	}
-	return g
-}
-
 func (b *groupMessageBuilder) build() *msg.Message {
 	sort.Slice(b.MessageSlices, func(i, j int) bool {
 		return b.MessageSlices[i].Content.GetPkgIndex() < b.MessageSlices[j].Content.GetPkgIndex()
@@ -704,6 +562,22 @@ func genLongTemplate(resId, brief string, ts int64) *message.ServiceElement {
 		ResId:   resId,
 		SubType: "Long",
 	}
+}
+
+func (c *QQClient) packOIDBPackage(cmd, serviceType int32, body []byte) []byte {
+	pkg := &oidb.OIDBSSOPkg{
+		Command:       cmd,
+		ServiceType:   serviceType,
+		Bodybuffer:    body,
+		ClientVersion: "Android " + c.version.SortVersionName,
+	}
+	r, _ := proto.Marshal(pkg)
+	return r
+}
+
+func (c *QQClient) packOIDBPackageProto(cmd, serviceType int32, msg proto.Message) []byte {
+	b, _ := proto.Marshal(msg)
+	return c.packOIDBPackage(cmd, serviceType, b)
 }
 
 func (c *QQClient) Error(msg string, args ...interface{}) {
